drivers:adc:ad7606: Fix build failures for platforms other than xilinx
Add compile time checks for AXI function calls

Move the device structure to ad7606.h

Remove static keyword in SPI reg read/write/write mask APIs as it is being used
outside the drivers.
